refactor: move asn1-related classes to subpackage for easier code navigation
